+++ src/pykeen/metrics/utils.py
@@ -7,6 +7,7 @@ from typing import ClassVar, Collection, Iterable, Optional
 
 import numpy as np
 from docdata import get_docdata
+from scipy import stats
 
 from ..utils import camel_to_snake
 
@@ -16,6 +17,8 @@ __all__ = [
     "stable_product",
     "weighted_mean_expectation",
     "weighted_mean_variance",
+    "weighted_harmonic_mean",
+    "weighted_median",
 ]
 
 
@@ -232,3 +235,47 @@ def stable_product(a: np.ndarray, is_log: bool = False) -> np.ndarray:
         sign = np.prod(np.copysign(np.ones_like(a), a))
         a = np.log(np.abs(a))
     return sign * np.exp(np.sum(a))
+
+
+def weighted_harmonic_mean(a: np.ndarray, weights: Optional[np.ndarray] = None) -> np.ndarray:
+    """
+    Calculate weighted harmonic mean.
+
+    :param a:
+        the array
+    :param weights:
+        the weight for individual array members
+
+    :return:
+        the weighted harmonic mean over the array
+
+    .. seealso::
+        https://en.wikipedia.org/wiki/Harmonic_mean#Weighted_harmonic_mean
+    """
+    if weights is None:
+        return stats.hmean(a)
+
+    # normalize weights
+    weights = weights.astype(float)
+    weights = weights / weights.sum()
+    # calculate weighted harmonic mean
+    return np.reciprocal(np.average(np.reciprocal(a.astype(float)), weights=weights))
+
+
+def weighted_median(a: np.ndarray, weights: Optional[np.ndarray] = None) -> np.ndarray:
+    """Calculate weighted median."""
+    if weights is None:
+        return np.median(a)
+
+    # calculate cdf
+    indices = np.argsort(a)
+    s_ranks = a[indices]
+    s_weights = weights[indices]
+    cdf = np.cumsum(s_weights)
+    cdf /= cdf[-1]
+    # determine value at p=0.5
+    idx = np.searchsorted(cdf, v=0.5)
+    # special case for exactly 0.5
+    if cdf[idx] == 0.5:
+        return s_ranks[idx : idx + 2].mean()
+    return s_ranks[idx]
